+++ Helios/Properties/AssemblyInfo.cs
@@ -45,5 +45,5 @@ using System.Windows;
 // You can specify all the values or you can default the Build and Revision Numbers 
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("1.4.2020.0530")]
-[assembly: AssemblyFileVersion("1.4.2020.0530")]
+[assembly: AssemblyVersion("1.4.2020.0603")]
+[assembly: AssemblyFileVersion("1.4.2020.0603")]